Update README.md
The calculator project is a web application that utilizes HTML, CSS, and JavaScript to create a functional and visually appealing calculator. Users can input numbers and perform basic arithmetic operations such as addition, subtraction, multiplication, and division. The JavaScript code handles the calculations and updates the display in real-time, while the CSS enhances the visual design and responsiveness of the calculator interface.